move helpers into new module, update log
