+++ tdrs-backend/tdpservice/data_files/views.py
@@ -20,7 +20,8 @@ from tdpservice.data_files.serializers import DataFileSerializer
 from tdpservice.data_files.models import DataFile
 from tdpservice.users.permissions import DataFilePermissions
 from tdpservice.scheduling import sftp_task
-from tdpservice.email.email_helper import send_data_submitted_email
+from tdpservice.email.helpers.data_file import send_data_submitted_email
+
 
 class DataFileFilter(filters.FilterSet):
     """Filters that can be applied to GET requests as query parameters."""
